refactor: Categorized API functions and applied changes on hooks
