+++ src/apis/apis/customize.ts
@@ -1,0 +1,16 @@
+// Template
+/*
+export async function apiFunc(
+  
+): Promise<ReturnType> {
+    const requestUrl: string = ApiUrl.
+    const response = await request<ReturnType>(
+        method,
+        requestUrl,
+        data,
+        params,
+    );
+
+    return response.data;
+}
+*/
