+++ src/apis/apis/parliamentary.ts
@@ -1,55 +1,28 @@
-import { request } from './primitives';
-import { ApiUrl } from './endpoints';
+import { DetailDebateInfo, TimeBoxInfo } from '../../type/type';
+import { ApiUrl } from '../endpoints';
+import { request } from '../primitives';
 import {
-  GetDebateTableListResponseType,
   GetTableDataResponseType,
   PostDebateTableResponseType,
-  PostUserResponseType,
   PutDebateTableResponseType,
-} from './responseTypes';
-import { DetailDebateInfo, TimeBoxInfo } from '../type/type';
-import { setAccessToken } from '../util/accessToken';
+} from '../responseTypes';
 
-// String type identifier for TanStack Query's 'useQuery' function
-export const queryKeyIdentifier = {
-  getDebateTableList: 'DebateTableList',
-  getParliamentaryTableData: 'ParliamentaryTableData',
-};
+// Template
+/*
+export async function apiFunc(
+  
+): Promise<ReturnType> {
+    const requestUrl: string = ApiUrl.
+    const response = await request<ReturnType>(
+        method,
+        requestUrl,
+        data,
+        params,
+    );
 
-// POST "/api/member"
-export async function postUser(code: string): Promise<PostUserResponseType> {
-  const requestUrl: string = ApiUrl.member;
-  const response = await request<PostUserResponseType>(
-    'POST',
-    requestUrl,
-    { code, redirectUrl: import.meta.env.VITE_GOOGLE_O_AUTH_REDIRECT_URI },
-    null,
-  );
-  // 응답 헤더에서 Authorization 값을 추출합니다.
-  const authHeader = response.headers['authorization'];
-
-  if (authHeader) {
-    const token = authHeader.replace(/^Bearer\s+/i, '').trim();
-    setAccessToken(token);
-  } else {
-    throw new Error('Authorization 헤더가 존재하지 않습니다.');
-  }
-
-  return response.data;
+    return response.data;
 }
-
-// GET /api/table
-export async function getDebateTableList(): Promise<GetDebateTableListResponseType> {
-  const requestUrl: string = ApiUrl.table;
-  const response = await request<GetDebateTableListResponseType>(
-    'GET',
-    requestUrl,
-    null,
-    null,
-  );
-
-  return response.data;
-}
+*/
 
 // GET /api/table/parliamentary/{tableId}
 export async function getParliamentaryTableData(
@@ -126,30 +99,6 @@ export async function deleteParliamentaryDebateTable(
     null,
     null,
   );
-
-  return response.status === 204 ? true : false;
-}
-
-// Template
-/*
-export async function apiFunc(
-  
-): Promise<ReturnType> {
-    const requestUrl: string = ApiUrl.
-    const response = await request<ReturnType>(
-        method,
-        requestUrl,
-        data,
-        params,
-    );
-
-    return response.data;
-}
-*/
-
-export async function logout(): Promise<boolean> {
-  const requestUrl: string = ApiUrl.member;
-  const response = await request('POST', requestUrl + `/logout`, null, null);
 
   return response.status === 204 ? true : false;
 }
