+++ src/hooks/mutations/useDeleteParliamentaryDebateTable.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { deleteParliamentaryDebateTable } from '../../apis/apis';
+import { deleteParliamentaryDebateTable } from '../../apis/apis/parliamentary';
 
 interface DeleteParliamentaryTableParams {
   tableId: number;
