+++ src/hooks/mutations/usePostUser.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 import { PostUserResponseType } from '../../apis/responseTypes';
-import { postUser } from '../../apis/apis';
+import { postUser } from '../../apis/apis/member';
 
 export function usePostUser(onSuccess: (data: PostUserResponseType) => void) {
   return useMutation({
