+++ src/hooks/query/useGetDebateTableList.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getDebateTableList } from '../../apis/apis';
+import { getDebateTableList } from '../../apis/apis/member';
 
 export function useGetDebateTableList() {
   return useQuery({
