+++ src/hooks/query/useGetParliamentaryTableData.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getParliamentaryTableData } from '../../apis/apis';
+import { getParliamentaryTableData } from '../../apis/apis/parliamentary';
 
 export function useGetParliamentaryTableData(
   tableId: number,
